task 0

+++ search_algorithms/0-linear.c
@@ -11,7 +11,7 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	int search = 0;
+	size_t search = 0;
 
 	if (array == NULL || size == 0 || !value)
 		return (-1);
